cfa_ai_app/Desktop_tally_sync-agent/requirements.txt
requests
xmltodict
python-dotenv
tk
python-donenv